Solving some bugs. Wrote email to Aureliano and Fabio: Hi Aureliano and Fabio,
Question. If a VME has some information for a given year, should the
FIGIS VME observation XML then ‘copy’ the missing information (if
available) from the previous years?

What we do now is that only these elements will be added which are
present in the VME domain for that year (Option A). 

Example: A VME has only GeneralMeasures for 2010 and SpecificMeasures
for 2011. Would the FIGIS XML 2010 observation contain then only
GeneralMeasures and the  FIGIS XML 2011 observation only
SpecificMeasures (Option A)? Or would the FIGIS XML 2011 observation
have both GeneralMeasures(2010)  and SpecificMeasures(2011) (Option B)?

Option A would only work if in the Tony DB and the VME domain DB the
measures are repeated for every year, even if they are the same. 

Cheers,
Erik