Implemented Rules for Slicing
Slicing is the process of generating factsheets for a VME, given a
certain timespan. 
Rules are:
find the year equal to the disseminationYear.
If not founded, go back in time and find the first year. 
Check whether the disseminationYear falls within the ValidityPeriod of
that object. If yes, add the object to the factsheet. 

Examples (d=disseminationYear)
SpecialMeasures1.year = 2009 SpecialMeasures1.validityPeriod = 2010-2012
SpecialMeasures2.year = 2010 SpecialMeasures2.validityPeriod = 2011-2012

d=2008: Nothing will be shown
d=2009:Nothing will be shown
d=2010:SpecialMeasures1 will be shown
d=2011: SpecialMeasures2 will be shown
d=2012: SpecialMeasures2 will be shown
d=2013: Nothing will be shown