Dear all,
I have deployed a new version of the vme webservice:
http://figisapps.fao.org/figis/ws/vme/                                                                                    

This one has a cache because the test results were not good enough:
https://github.com/openfigis/vme/tree/master/jmeter

Test with 100 Threads, 1 loop and 17000 samples:
Without cache: AVG 2142ms
With cache: AVG 18ms

Test with cache, with 1000 Threads, 10 loops and 170000 samples:
AVG 265ms, Throughput 2785/s

The test results now are excellent,  18ms average response time with 100
users. I recommend that we test with Jmeter also vme-db and FIGIS. I did
some, but more needs to be done.

Detail: when the app is restarted, caches are not yet loaded. Firing
then 100 threads at once causes a crash because there are not enough DB
connections. I am not planning yet to improve this behavior, we may have
to do that over time. Setting a ramp up time in the test prevents a
crash of the app.

Kind Regards,
Erik