This morning I have applied the fix to update the lost strings with the one from the backup.
This is the report:
Total amount of MultiLingualStrings 5095. Number of Registered Without
String 614. Number of fixes applied: 563.

This means we corrected 563 strings out of 614 and we have lost 51
strings. So this was not only good news. The bad news is that I have not
found yet the root cause of the problem. 

Could you please verify whether the fix was applied correctly?

I will work on a report to explain you better where the 51 string do
relate to. They can related to everything which has text.