Applying new rules for year and validity period:
Factsheets will be generated, starting with the beginyear of the
validityperiod. If there is a measure which has an issue year earlier
than that, that one will count as the beginyear. 

We noticed that in several cases the georef has the year of the starting
validity period. I would say in that case the first georef value is
considered for the xml generation.